organization of some components
