+++ src/components/content.jsx
@@ -9,8 +9,8 @@ import Sleep from './sleep/sleep'
 import Diaper from './diaper/diaper'
 import Feeding from './feeding/feeding'
 import Children from './children/children'
-import Logout from './logout'
-import LoginForm from './loginform'
+import Logout from './authentication/logout'
+import LoginForm from './authentication/loginform'
 
 
 class Content extends React.Component {
